# Add neon SIMD alpha blending blitter