I've added neon SIMD code for alpha blending blitter. It runs **~3.5x to 6x faster** than scalar code.

Neon code is enabled only for 64-bit arm. Supporting 32-bit arm would require a extra code (because of `vqtbl1q_u8`). 
For 64-bit arm neon is always present, so new code is enabled unconditionally.
Neon calculates exactly same byte values as scalar code.

To measure performance I called SDL_BlitSurface for 256x256 ABGR & ARGB surfaces 1000x in a loop.
Here are timings for old scalar -> new neon code improvement.

Raspberry Pi 5, gcc 14.1.1
```
ABGR->ABGR: 244.7 msec -> 63.1 msec = 3.88x faster
ABGR->ARGB: 376.4 msec -> 63.1 msec = 5.97x faster
```

Apple M1, clang 15.0.0
```
ABGR->ABGR: 76.2 msec -> 20.8 msec = 3.66x faster
ABGR->ARGB: 94.2 msec -> 20.8 msec = 4.53x faster
```

I have run `testautomation --filter Blending` on both devices to check that hashes of blit results matches expected values.

Code also successfully compiles with MSVC for arm64 target, but I have no Windows arm64 device that I can test it on.
(cmake build requires minor patching in CMakeLists.txt otherwise it always sets SDL_DISABLE_NEON for MSVC)