@slouken 2024-08-29
Nice work! Did you test odd widths to make sure the remainder code is correct?