@mmozeiko 2024-08-29
Yes, non-4 multiple widths work fine. The `testautomation_blit.c` tests all kinds of sizes.