@madebr 2024-08-29
About CMake failing to enable neon, does the following test fail on your arm64 system:
https://github.com/libsdl-org/SDL/blob/d9b3b9ad91f0883b7840a589ecbc7ed27f8a7852/CMakeLists.txt#L853-L863

You can find the outputs of cmake configuration tests in `<build-directory>/CMakeFiles/CMakeConfigureLog.yaml`